fix(lint): clear 3 eslint warnings
- Extract labToRgb into scene-browser/utils/color-convert to satisfy
  react-refresh/only-export-components on scene-palette-swatches.tsx
- Drop now-unused no-var disable comments in media-library and
  scene-browser global store declarations
